requirements.txt
Flask==2.0.3
cassandra-driver==3.25.0